azs - DL - update fragility and consequence figure generator for PBE
Now that Pelicun3 is not part of the backend, the figures need to be
generated using a dynamic location for the underlying data. This is
handled by adding a 'query' option to this script that returns the
location of default DL model data files.

To avoid generating figures multiple times for the same underlying data,
the data files are copied to the output folder. Whenever a new request
comes to generate figures, the new source data is compared to the
existing one in the output folder. The figures are only generated if
there is difference between the existing and new data. The comparison
is handled by the new check_diff function.