Extend rWHALE to only look for requested results.
rWHALE used to blindly try to open every type of result file, collect
the information and then check what to save in the output. Instead, this
commit adds code that checks what outputs are requested and utilizes the
existing argument of aggregate_results to ask for only a subset of
output types if that's what the user needs.